requirements.txt
pytesseract
opencv-python
numpy
pdf2image
python-dotenv
openai 